Use async in the handlers (#10)
Currently no async iterators so I had to remove the queue iterator. This more than doubles the throughput of the stats endpoint.

Had to update to Actix Web 4 beta to get access to the Tokio 1.0 runtime required by the newer Redis 0.19+. A little annoying but it seems like everything kept working.